메세지 카운트
업데이트 :
- 메세지 사용횟수를 저장해 씬이 바뀌거나(스토리 다시보기 같은거) 게임을 재시작 해도 메세지 카운트가 초기화 되지 않습니다

+++ Assets/02. Scripts/Managers/FileManager.cs
@@ -58,6 +58,7 @@ public class FileManager : MonoBehaviour
         public bool LockFirst;
         public bool LockLast;
         public bool LockItem;
+        public int msgCount;
     }
 
     [System.Serializable]
@@ -131,7 +132,7 @@ public class FileManager : MonoBehaviour
     // ========================================================================
     // [Part 1] 워드이터 게임 데이터 (레벨, 정답, 히스토리)
     // ========================================================================
-    public void SaveWordEaterInfo(int le, string an, string hi, List<string> RR , string id , string RRL,bool LLen,bool LF, bool LLast , bool LI)
+    public void SaveWordEaterInfo(int le, string an, string hi, List<string> RR , string id , string RRL,bool LLen,bool LF, bool LLast , bool LI,int mc)
     {
         WordEaterData data = new WordEaterData
         {
@@ -149,7 +150,8 @@ public class FileManager : MonoBehaviour
             LockLength = LLen,
             LockFirst = LF,
             LockLast = LLast,
-            LockItem = LI
+            LockItem = LI,
+            msgCount = mc
         };
 
         string json = JsonUtility.ToJson(data, true);
@@ -237,6 +239,7 @@ public class FileManager : MonoBehaviour
             gamemanager.isFirst = data.LockFirst;
             gamemanager.isLast = data.LockLast;
             gamemanager.isChoseongItem = data.LockItem;
+            gamemanager.messageCount = data.msgCount;
         }
 
         // [추가] 로드된 키 데이터를 임시 보관 또는 즉시 적용
@@ -276,6 +279,16 @@ public class FileManager : MonoBehaviour
         string json = File.ReadAllText(WordEaterPath);
         WordEaterData data = JsonUtility.FromJson<WordEaterData>(json);
         data.Relevant = Rel;
+
+        File.WriteAllText(WordEaterPath, JsonUtility.ToJson(data, true));
+    }
+
+    public void saveMsgCount(int msgC) {
+        if (!File.Exists(WordEaterPath)) return;
+
+        string json = File.ReadAllText(WordEaterPath);
+        WordEaterData data = JsonUtility.FromJson<WordEaterData>(json);
+        data.msgCount = msgC;
 
         File.WriteAllText(WordEaterPath, JsonUtility.ToJson(data, true));
     }
